Refactor bzl_lock_generator annotation handling
Introduces an internal `PackageAnnotations` struct that collects things like
build dependencies, ignored dependencies, etc. prior to constructing PackageTargets.

This cleans up the prior messiness that existed due to PackageTarget objects being
modified after construction to apply these options.

The primary driver behind this change was to push the handling of ignored dependencies
to a deeper layer, so that we don't generate wheel/sdist repos and other artifacts for
packages that are completely ignored.