Add `post_extraction_command`
This flag is useful when one needs to do some ad-hoc operations on the
wheel files. For example, our use case pulls in a lot of third party
dependencies and some of them may have name conflict each other and we
sometimes need to remove or patch `__init__.py` files in order for
conflicting dependencies to coexist.

I am open to other suggestions or work arounds. But adding the ability
to do ad-hoc patches for each wheel_library seems to be the simplest way
to achieve this.