Use an absolute path to C/C++ compilers for better compatibility
[`toolchains_llvm`][1] only triggers its logic to locate files relative
to the compiler wrapper when called with an absolute path, for example.
Otherwise it assumes it's being called from the bazel execroot, which is
not true in this case.

`toolchains_llvm` already has similar logic as `wheel_builder` for
mapping paths, which could be used by just calling it with an absolute
path instead of using `wheel_builder`'s wrapper, but I think it's easier
to use the `wheel_builder` version instead of trying to detect whether
the toolchain has this logic.

[1]: https://github.com/bazel-contrib/toolchains_llvm/blob/1d685a99db5f1e6523c392e7dfaced583062cbc4/toolchain/cc_wrapper.sh.tpl#L37