Add support for exposing C/C++ dependencies
Some Python libraries (`numpy` is the most common example) support C/C++
code depending on them directly. To build this C/C++ code with Bazel, we
need to expose the header files.

Typically C/C++ code depending on headers from Python packages is going
to be imported as other Python packages, so it does not link directly to
any definitions for the APIs declared in the header files it uses.
`cc_srcs` could be added alongside `cc_hdrs_globs` to enable this if
there's a use case in the future.